Fix lag, change button to switch
Lag was caused by adding a timeline for every bubble created. Changed it so the bubbles spawn on a single timeline as they just stagger into view anyway.

Replaced the if-else start button with a switch so that it can go through a trigger warning before bombarding the user with messages.

Initial work into draggable functionality for the bubbles. Not sure why it's not working by default though.