Update main.js
Released the loop creating the bubbles from its own function so that I could attach a timeline to those bubbles that can then be played/restarted on button press.

Previously, bubble creation + their positioning would be done asynchronously, which meant the timeline had to be done so as well, meaning I could not play/pause the timeline because it couldn't attach to a target that didn't exist yet.